+++ crypto/x509/x509.go
@@ -4,6 +4,7 @@ package x509
 #cgo pkg-config: openssl
 #include <openssl/pem.h>
 #include <openssl/x509.h>
+#include <openssl/ssl.h>
 
 #define d2i_X509_f 0
 #define d2i_X509_REQ_f 1
@@ -85,6 +86,10 @@ type Certificate struct {
 	DNSNames    []string
 
 	//TODO(runcom): add more pub fields
+}
+
+func (c *Certificate) GetX509() *C.X509 {
+	return c.x509
 }
 
 // ParseCertificate parses a single certificate from the given ASN.1 DER data.
@@ -257,4 +262,63 @@ func EncryptPEMBlock(rand io.Reader, blockType string, data, password []byte, al
 // IsEncryptedPEMBlock returns if the PEM block is password encrypted.
 func IsEncryptedPEMBlock(b *pem.Block) bool {
 	return x509.IsEncryptedPEMBlock(b)
+}
+
+// taken from old, top ns, gossl
+
+//Export an OpenSSL X509 to a DER buffer
+func (self *Certificate) DumpDERCertificate() ([]byte, error) {
+	bio := C.BIO_new(C.BIO_s_mem())
+	defer C.BIO_free(bio)
+	ret := C.i2d_X509_bio(bio, self.x509)
+	if ret == 0 {
+		return nil, errors.New("problem dumping certificate")
+	}
+	var temp *C.char
+	buf_len := C.BIO_ctrl(bio, C.BIO_CTRL_INFO, 0, unsafe.Pointer(&temp))
+	return C.GoBytes(unsafe.Pointer(temp), C.int(buf_len)), nil
+}
+
+//Helper function that calls encoding/pem to convert DER to PEM
+func ParseCertificatePEM(pemData []byte) (*Certificate, error) {
+	length := C.int(len(pemData))
+	buffer := unsafe.Pointer(&pemData[0])
+	bio := C.BIO_new_mem_buf(buffer, length)
+	cert := C.PEM_read_bio_X509(bio, nil, nil, nil)
+	if cert == nil {
+		return nil, errors.New("problem loading certificate" + sslerr.SSLErrorMessage().String())
+	}
+	return &Certificate{x509: cert}, nil
+
+}
+
+type X509Store struct {
+	Store *C.X509_STORE
+}
+
+func NewX509Store(ctx unsafe.Pointer) *X509Store {
+	cctx := (*C.SSL_CTX)(ctx)
+	return &X509Store{Store: C.SSL_CTX_get_cert_store(cctx)}
+}
+
+func (self *X509Store) SetDepth(depth int) int {
+	return int(C.X509_STORE_set_depth(self.Store, C.int(depth)))
+}
+func (self *X509Store) AddCert(cert *Certificate) int {
+	return int(C.X509_STORE_add_cert(self.Store, cert.x509))
+}
+
+type X509Name struct {
+	Name *C.X509_NAME
+}
+
+func (self *X509Name) Print() ([]byte, error) {
+	bio := C.BIO_new(C.BIO_s_mem())
+	defer C.BIO_free(bio)
+	//TODO check for error here
+	C.X509_NAME_print_ex(bio, self.Name, 0, C.XN_FLAG_MULTILINE)
+	var temp *C.char
+	buf_len := C.BIO_ctrl(bio, C.BIO_CTRL_INFO, 0, unsafe.Pointer(&temp))
+	defer C.free(unsafe.Pointer(temp))
+	return C.GoBytes(unsafe.Pointer(temp), C.int(buf_len)), nil
 }
